% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{aw_scores}
\alias{aw_scores}
\title{Compute AIPW/doubly robust scores.}
\usage{
aw_scores(yobs, ws, balwts, K, mu_hat = NULL)
}
\arguments{
\item{yobs}{observed rewards, a numeric vector of length A.}

\item{ws}{pulled arms, an integer vector of length A.}

\item{balwts}{inverse probability score 1[W_t=w]/e_t(w) of pulling arms, a numeric matrix of
shape [A, K].}

\item{K}{number of arms, an integer.}

\item{mu_hat}{plug-in estimator of arm outcomes, a numeric matrix of shape [A, K], or NULL.}
}
\value{
scores AIPW scores, a numeric matrix of shape [A, K].
}
\description{
Compute AIPW/doubly robust scores based on observed rewards, pulled arms, and inverse
probability scores. If mu_hat is provided, compute AIPW scores, otherwise compute IPW scores.
}
\examples{
aw_scores(yobs = c(0.5, 1, 0, 1.5), ws = c(1, 2, 2, 3),
          balwts = matrix(c(0.5, 2, 1, 0.5, 1, 1.5, 0.5, 1.5, 2, 1.5, 0.5, 1), ncol = 3),
          K = 3,
          mu_hat = matrix(c(0.5, 0.8, 0.6, 0.3, 0.9, 0.2, 0.5, 0.7, 0.4, 0.8, 0.2, 0.6), ncol = 3))

}
