% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\docType{data}
\name{balwts}
\alias{balwts}
\title{balwts Function}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1000 rows and 4 columns.
}
\usage{
balwts
}
\arguments{
\item{ws}{a vector of length A containing the actions taken at each time step}

\item{probs}{a matrix of shape [A, K] or [A, A, K] containing the true probabilities of each arm being chosen at each time step}
}
\value{
A matrix of shape [A, K] or [A, A, K] containing the inverse probability score of pulling arms
}
\description{
The balwts function calculates the inverse probability score of pulling arms, given the actions taken and the true probabilities of each arm being chosen.
}
\examples{
A <- 5
K <- 3
ws <- sample(1:K, A, replace = TRUE)
probs <- matrix(runif(A * K), nrow = A, ncol = K)
balwts <- balwts(ws, probs)

}
\keyword{datasets}
