% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{calculate_mu_hat}
\alias{calculate_mu_hat}
\title{calculate_mu_hat Function}
\usage{
calculate_mu_hat(results)
}
\arguments{
\item{results}{a list containing the results of a bandit model, including the fitted model and the covariate matrix xs}
}
\value{
A vector containing the estimated expected value of the reward for each arm
}
\description{
The calculate_mu_hat function calculates the estimated expected value of the reward for each arm, given the results of a bandit model.
}
\examples{
A <- 200
K <- p <- 3
xs <- matrix(runif(A * p), nrow = A, ncol = p)
ys <- matrix(rbinom(A * K, 1, 0.5), nrow = A, ncol = K)
batch_sizes <- c(100,100)
results <- run_experiment(ys = ys, floor_start = 5, floor_decay = 0.9, batch_sizes = batch_sizes, xs = xs)
mu_hat <- calculate_mu_hat(results)

}
