% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{.check_A}
\alias{.check_A}
\title{Check Number of Observations for Inference}
\usage{
.check_A(A)
}
\arguments{
\item{A}{An integer representing the number of observations.}
}
\value{
Returns NULL if the number of observations is valid; otherwise, throws an error.
}
\description{
This function checks if the number of observations is greater than 1, which is required for conducting inference.
}
