% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{.check_shape}
\alias{.check_shape}
\title{Check Shape Compatibility of Probability Objects}
\usage{
.check_shape(gammahats, contextual_probs)
}
\arguments{
\item{gammahats}{A matrix representing estimates.}

\item{contextual_probs}{An object representing probabilities, either a matrix (non-contextual) or an array (contextual).}
}
\value{
Returns NULL if the shapes and types are valid; otherwise, throws an error.
}
\description{
This function checks the dimensional compatibility of `gammahats` and `contextual_probs`. It validates the dimensions and types based on whether the probability objects are contextual or non-contextual.
}
