% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{impose_floor}
\alias{impose_floor}
\title{impose_floor Function}
\usage{
impose_floor(a, amin)
}
\arguments{
\item{a}{a numeric vector to impose a floor on}

\item{amin}{a numeric value indicating the minimum allowed value}
}
\value{
A numeric vector with the same length as a, with the floor imposed on its elements
}
\description{
The impose_floor function imposes a floor on the given array a, ensuring that its elements are greater than or equal to amin.
}
\examples{
a <- c(0.25, 0.25, 0.25, 0.25)
imposed_a <- impose_floor(a=a, amin=0.1)

}
