% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{LinTSModel}
\alias{LinTSModel}
\title{LinTSModel Function}
\usage{
LinTSModel(K, p, floor_start, floor_decay, num_mc = 100, is_contextual = TRUE)
}
\arguments{
\item{K}{the number of arms in the bandit problem}

\item{p}{the dimension of the contextual vector, if is_contextual is set to TRUE. Otherwise, p is ignored}

\item{floor_start}{the initial floor value for the exploration parameter}

\item{floor_decay}{the decay rate of the floor value for the exploration parameter}

\item{num_mc}{the number of Monte Carlo simulations used to approximate the expected reward}

\item{is_contextual}{a logical value indicating whether the problem is contextual or not}
}
\value{
A list containing the parameters of the LinTSModel
}
\description{
The LinTSModel function creates a linear Thompson Sampling model for multi-armed bandit problems.
}
\examples{
model <- LinTSModel(K=5, p=3, floor_start=1, floor_decay=0.9, num_mc=100, is_contextual=TRUE)

}
