% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{plot_cumulative_assignment}
\alias{plot_cumulative_assignment}
\title{plot_cumulative_assignment Function}
\usage{
plot_cumulative_assignment(results, batch_sizes)
}
\arguments{
\item{results}{a list containing the results of the experiment, including the actions taken (ws) and the true probabilities of each arm being chosen (probs)}

\item{batch_sizes}{a vector of batch sizes used in the experiment}
}
\value{
A plot of the cumulative assignment graph
}
\description{
The plot_cumulative_assignment function generates a plot of the cumulative assignment graph for every arm and every batch size.
}
\examples{
set.seed(123)
A <- 1000
K <- 4
xs <- matrix(runif(A * K), nrow = A, ncol = K)
ys <- matrix(rbinom(A * K, 1, 0.5), nrow = A, ncol = K)
batch_sizes <- c(250, 250, 250, 250)
results <- run_experiment(ys = ys,
                          floor_start = 5,
                          floor_decay = 0.9,
                          batch_sizes = batch_sizes,
                          xs = xs)
plot_cumulative_assignment(results, batch_sizes)

}
