% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{ridge_init}
\alias{ridge_init}
\title{Ridge Regression Initialization for Arm Expected Rewards}
\usage{
ridge_init(p, K)
}
\arguments{
\item{p}{Number of covariates.}

\item{K}{Number of arms.}
}
\value{
A list containing initialized matrices `R_A`, `R_Ainv`, `b`, and `theta` for each arm.
}
\description{
Initializes matrices needed for ridge regression to estimate the expected rewards of different arms.
}
\examples{
p <- 3
K <- 5
init <- ridge_init(p, K)

}
