% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{run_experiment}
\alias{run_experiment}
\title{run_experiment Function}
\usage{
run_experiment(
  ys,
  floor_start,
  floor_decay,
  batch_sizes,
  xs = NULL,
  balanced = NULL
)
}
\arguments{
\item{ys}{a matrix of potential outcomes of shape [A, K], where A is the number of observations and K is the number of arms}

\item{floor_start}{a numeric value indicating the starting value of the floor}

\item{floor_decay}{a numeric value indicating the decay rate of the floor}

\item{batch_sizes}{a vector of integers indicating the size of each batch}

\item{xs}{an optional matrix of contextual vectors for the observations, if the LinTSModel is contextual}

\item{balanced}{an optional boolean value indicating whether to balance the batches}
}
\value{
A list containing the pulled arms (ws), observed rewards (yobs), assignment probabilities (probs), and the fitted bandit model (fitted_bandit_model)
}
\description{
The run_experiment function runs a LinTS or non-contextual TS bandit experiment, given potential outcomes and covariates.
}
\examples{
ys <- matrix(rbinom(1000, 1, 0.5), ncol=5)
xs <- matrix(rnorm(300), ncol=3)
batch_sizes <- c(100, 200, 700)
result <- run_experiment(ys=ys,
                          floor_start=1,
                          floor_decay=0.9,
                          batch_sizes=batch_sizes,
                          xs=xs,
                          balanced=TRUE)

}
