% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{stick_breaking}
\alias{stick_breaking}
\title{Stick breaking function.}
\usage{
stick_breaking(Z)
}
\arguments{
\item{Z}{Numeric array. Input array, shape \code{[A, K]}. Must not contain NA values.}
}
\value{
Numeric array. Stick breaking weights, shape \code{[A, K]}. Must not contain NA values.
}
\description{
Implements the stick breaking algorithm for calculating weights in the stable-var scheme.
}
\examples{
set.seed(123)
Z <- array(runif(10), dim = c(2, 5))
stick_breaking(Z)

}
