% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{twopoint_stable_var_ratio}
\alias{twopoint_stable_var_ratio}
\title{Calculate the allocation ratio for a two-point stable-variance bandit}
\usage{
twopoint_stable_var_ratio(e, alpha)
}
\arguments{
\item{e}{the empirical mean of the bad arm}

\item{alpha}{the discount parameter}
}
\value{
the allocation ratio lambda
}
\description{
This function calculates the allocation ratio for a two-point stable-variance bandit, given the empirical mean and the discount parameter alpha.
}
\examples{
# Calculate the allocation ratio for a two-point stable-variance bandit with e=0.1 and alpha=0.5
twopoint_stable_var_ratio(0.1, 0.5)

}
