% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{update_thompson}
\alias{update_thompson}
\title{update_thompson Function}
\usage{
update_thompson(ws, yobs, model, xs = NULL, ps = NULL, balanced = FALSE)
}
\arguments{
\item{ws}{a vector of integers indicating which arm was chosen for each observation}

\item{yobs}{a vector of numeric values representing the observed rewards}

\item{model}{a list containing the parameters of the LinTSModel}

\item{xs}{an optional matrix of contextual vectors for the observations, if the LinTSModel is contextual}

\item{ps}{an optional matrix of probabilities of selecting each arm for each observation, if the LinTSModel is balanced}

\item{balanced}{a logical value indicating whether to use balanced Thompson Sampling or not}
}
\value{
A list containing the updated parameters of the LinTSModel
}
\description{
The update_thompson function updates the parameters of a linear Thompson Sampling model for multi-armed bandit problems based on new observations.
}
\examples{
model <- LinTSModel(K=5, p=3, floor_start=1, floor_decay=0.9, num_mc=100, is_contextual=TRUE)
model <- update_thompson(ws=c(3,2,1), yobs=c(1,0,1), model=model, xs=matrix(rnorm(9), ncol=3))

}
