% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{aw_var}
\alias{aw_var}
\title{Variance of policy value estimator via non-contextual adaptive weighting.}
\usage{
aw_var(scores, estimate, policy, evalwts = NULL)
}
\arguments{
\item{scores}{AIPW score, a numeric matrix of shape [A, K].}

\item{estimate}{policy value estimate, a numeric scalar.}

\item{policy}{policy matrix pi(X_i, w), a numeric matrix of shape [A, K].}

\item{evalwts}{non-contextual adaptive weights h_i, a numeric vector of length A, or NULL.}
}
\value{
variance of policy value estimate, a numeric scalar.
}
\description{
Compute the variance of a policy value estimate based on AIPW scores, a policy matrix, and
non-contextual adaptive weights.
}
\examples{
scores <- matrix(c(0.5, 0.8, 0.6, 0.3, 0.9, 0.2, 0.5, 0.7, 0.4, 0.8, 0.2, 0.6), ncol = 3)
policy <- matrix(c(0.2, 0.3, 0.5, 0.6, 0.1, 0.3, 0.4, 0.5, 0.1, 0.2, 0.7, 0.1), ncol = 3)
estimate <- aw_estimate(scores = scores, policy = policy, evalwts = c(0.5, 1, 0.5, 1.5))
aw_var(scores = scores, estimate = estimate, policy = policy, evalwts = c(0.5, 1, 0.5, 1.5))

}
