% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{calculate_balwts}
\alias{calculate_balwts}
\title{calculate_balwts Function}
\usage{
calculate_balwts(ws, probs)
}
\arguments{
\item{ws}{a vector of length A containing the actions taken at each time step}

\item{probs}{a matrix of shape [A, K] or [A, A, K] containing the true probabilities of each arm being chosen at each time step}
}
\value{
A matrix of shape [A, K] or [A, A, K] containing the inverse probability score of pulling arms
}
\description{
The calculate_balwts function calculates the inverse probability score of pulling arms, given the actions taken and the true probabilities of each arm being chosen.
}
\examples{
A <- 5
K <- 3
ws <- sample(1:K, A, replace = TRUE)
probs <- matrix(runif(A * K), nrow = A, ncol = K)
balwts <- calculate_balwts(ws, probs)

}
