% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{draw_thompson}
\alias{draw_thompson}
\title{draw_thompson Function}
\usage{
draw_thompson(model, start, end, xs = NULL)
}
\arguments{
\item{model}{a list containing the parameters of the LinTSModel}

\item{start}{an integer indicating the starting index of the observations to make the draw}

\item{end}{an integer indicating the ending index of the observations to make the draw}

\item{xs}{an optional matrix of contextual vectors for the observations, if the LinTSModel is contextual}
}
\value{
A list containing the drawn arms (w) and their corresponding probabilities (ps)
}
\description{
The draw_thompson function draws arms from a LinTS or non-contextual TS agent for multi-armed bandit problems.
}
\examples{
model <- LinTSModel(K=5, p=3, floor_start=1, floor_decay=0.9, num_mc=100, is_contextual=TRUE)
draws <- draw_thompson(model=model, start=1, end=10, xs=matrix(rnorm(30), ncol=3))

}
