% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{generate_bandit_data}
\alias{generate_bandit_data}
\title{generate_bandit_data Function}
\usage{
generate_bandit_data(X = NULL, y = NULL, noise_std = 1, signal_strength = 1)
}
\arguments{
\item{X}{an optional matrix of covariates of shape [N, p], where N is the number of observations and p is the number of features}

\item{y}{an optional vector of labels of length N}

\item{noise_std}{a numeric value indicating the standard deviation of the noise added to the potential outcomes}

\item{signal_strength}{a numeric value indicating the strength of the signal in the potential outcomes}
}
\value{
A list containing the generated data (xs, ys, muxs, A, p, K) and the true class probabilities (mus)
}
\description{
The generate_bandit_data function generates covariates and potential outcomes from a classification dataset.
}
\examples{
data <- generate_bandit_data(X=iris[,1:4], y=iris[,5], noise_std=0.1, signal_strength=1.0)
xs <- data$data$xs
ys <- data$data$ys

}
