% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{ifelse_clip}
\alias{ifelse_clip}
\title{Clip lamb values between a minimum x and maximum y}
\usage{
ifelse_clip(lamb, x, y)
}
\arguments{
\item{lamb}{a numeric vector to be clipped}

\item{x}{the lower bound of the clip range}

\item{y}{the upper bound of the clip range}
}
\value{
a clipped numeric vector
}
\description{
This function clips a numeric vector between two values.
}
\examples{
lamb <- c(1, 2, 3, 4, 5)
ifelse_clip(lamb, 2, 4)

}
