% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{output_estimates}
\alias{output_estimates}
\title{Calculate treatment effect estimates using AIPW with various weighting schemes}
\usage{
output_estimates(
  policy0 = NULL,
  policy1,
  contrasts = "combined",
  gammahat,
  contextual_probs,
  uniform = TRUE,
  non_contextual_minvar = TRUE,
  contextual_minvar = TRUE,
  non_contextual_stablevar = TRUE,
  contextual_stablevar = TRUE,
  non_contextual_twopoint = TRUE
)
}
\arguments{
\item{policy0}{A * K control policy matrix for contrast evaluation, with probabilities under control. When `policy0 = NULL`, the function is estimating the value \eqn{Q(w)} of a single arm w. When `policy0` doesn't equal to `NULL`, the function is estimating treatment effects of policies as compared to control \eqn{\Delta(w_1, w_2)}, using the difference in AIPW scores as the unbiased scoring rule for \eqn{\Delta (w_1, w_2)}.}

\item{policy1}{list of A * K counterfactual treatment policy matrices for evaluation, with assignment probabilities under each policy.}

\item{contrasts}{Define the approach to estimate treatment effects. `combined` indicates the first approach -- the difference in AIPW scores as the unbiased scoring rule for \eqn{\Delta (w_1, w_2)}; `separate` indicates the second approach -- \eqn{\hat \Delta (w_1, w_2) = \hat Q (w_1) - \hat Q (w_2)}.}

\item{gammahat}{Scores matrix.}

\item{contextual_probs}{A * A * K matrix for contextual probabilities, with dimensions representing time, contexts, treatment arms.}

\item{uniform}{Logical, estimate uniform weights.}

\item{non_contextual_minvar}{Logical, estimate non-contextual minvar weights.}

\item{contextual_minvar}{Logical, estimate contextual minvar weights.}

\item{non_contextual_stablevar}{Logical, estimate non-contextual stablevar weights.}

\item{contextual_stablevar}{Logical, estimate contextual stablevar weights.}

\item{non_contextual_twopoint}{The non_contextual_twopoint parameter.}
}
\value{
A list of treatment effect estimates under different weighting schemes.
}
\description{
This function calculates treatment effect estimates using augmented inverse probability weighting (AIPW) with various weighting schemes. The function estimates the value of a single arm or treatment effects of policies as compared to the control, using the difference in AIPW scores as the unbiased scoring rule or the difference in means between two policies. The function provides estimates under various weighting schemes, including uniform, non-contextual minvar, contextual minvar, non-contextual stablevar, contextual stablevar, and non-contextual two-point. The function is based on the algorithm developed in Zhan et al. (2021).
}
\examples{
# Generate simulated data
data <- generate_data()
# Estimate treatment effects using AIPW with various weighting schemes
estimates <- output_estimates(policy1 = data$policy1,
                              gammahat = data$gammahat,
                              contextual_probs = data$contextual_probs)
}
