% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{plot_cumulative_assignment}
\alias{plot_cumulative_assignment}
\title{plot_cumulative_assignment Function}
\usage{
plot_cumulative_assignment(results, batch_sizes)
}
\arguments{
\item{results}{a list containing the results of the experiment, including the actions taken (ws) and the true probabilities of each arm being chosen (probs)}

\item{batch_sizes}{a vector of batch sizes used in the experiment}
}
\value{
A plot of the cumulative assignment graph
}
\description{
The plot_cumulative_assignment function generates a plot of the cumulative assignment graph for every arm and every batch size.
}
\examples{
A <- 5
K <- 3
xs <- matrix(runif(A * K), nrow = A, ncol = K)
ys <- matrix(rbinom(A * K, 1, 0.5), nrow = A, ncol = K)
model <- bandit_lin_ucb(xs, ys)
experiment <- run_experiment(model, batch_sizes = c(10, 20, 30))
plot_cumulative_assignment(experiment$results, experiment$batch_sizes)

}
