% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{ridge_muhat_lfo_pai}
\alias{ridge_muhat_lfo_pai}
\title{Plug-in Estimates for Arm Expected Rewards Using Ridge Regression}
\usage{
ridge_muhat_lfo_pai(xs, ws, yobs, K, batch_sizes, alpha = 1)
}
\arguments{
\item{xs}{Matrix of observed covariates of shape [A, p].}

\item{ws}{Vector of pulled arm indices at each time `t` of shape [A].}

\item{yobs}{Vector of observed outcomes of shape [A].}

\item{K}{Number of arms.}

\item{batch_sizes}{Vector indicating sizes of batches in which data is processed.}

\item{alpha}{The ridge regression regularization parameter (default is 1).}
}
\value{
A 3D array containing the expected reward estimates for each arm and each time `t`, of shape [A, A, K].
}
\description{
Computes plug-in estimates of arm expected rewards based on provided data.
}
