% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{ridge_update}
\alias{ridge_update}
\title{Ridge Regression Update for Arm Expected Rewards}
\usage{
ridge_update(R_A, b, xs, t, yobs, alpha)
}
\arguments{
\item{R_A}{The current matrix A for an arm.}

\item{b}{The current vector b for an arm.}

\item{xs}{Matrix of observed covariates.}

\item{t}{The current time or instance.}

\item{yobs}{The observed outcome for the current observation.}

\item{alpha}{The ridge regression regularization parameter (default is 1).}
}
\value{
A list containing updated matrices `R_A`, `R_Ainv`, `b`, and `theta`.
}
\description{
Given previous matrices and a new observation, updates the matrices for ridge regression.
}
