% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_utils.R
\name{simple_tree_data}
\alias{simple_tree_data}
\title{simple_tree_data function}
\usage{
simple_tree_data(
  A,
  K = 5,
  p = 10,
  noise_std = 1,
  split = 1.676,
  signal_strength = 1,
  seed = NULL,
  noise_form = "normal"
)
}
\arguments{
\item{A}{Integer. Number of observations in the dataset.}

\item{K}{Integer. Number of treatment arms.}

\item{p}{Integer. Number of covariates.}

\item{noise_std}{Numeric. Standard deviation of the noise added to the potential outcomes.}

\item{split}{Numeric. Split point for creating treatment groups based on the covariates.}

\item{signal_strength}{Numeric. Strength of the signal in the potential outcomes.}

\item{seed}{Integer. Seed value for reproducibility.}

\item{noise_form}{Character. Distribution of the noise added to the potential outcomes.
Can be either "normal" or "uniform".}
}
\value{
A list containing the generated dataset and the true potential outcome means.
  The dataset includes the following components:
  \itemize{
    \item \code{xs}: Covariate matrix.
    \item \code{ys}: Potential outcome matrix.
    \item \code{muxs}: True treatment assignment matrix.
    \item \code{wxs}: Binary treatment indicator matrix.
  }
  The true potential outcome means are stored in the \code{mus} component.
}
\description{
Generates covariates and potential outcomes of a synthetic dataset for a simple tree model.
}
\examples{
A <- 1000
K <- 4     # Number of treatment arms
p <- 10    # Number of covariates
noise_std <- 1.0      # Standard deviation of noise
split <- 1.676       # Split value for generating binary assignments
signal_strength <- 1.0  # Signal strength multiplier
seed <- 123           # Seed for reproducibility
noise_form <- 'normal'  # Noise form (either 'normal' or 'uniform')

}
