% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_utils.R
\name{stick_breaking}
\alias{stick_breaking}
\title{stick_breaking function}
\usage{
stick_breaking(Z)
}
\arguments{
\item{Z}{input array of shape [A, K]}
}
\value{
weights stick_breaking weights of shape [A, K]
}
\description{
This function implements the stick breaking algorithm for calculating weights in the stable-var scheme.
}
\examples{
set.seed(123)
Z <- array(runif(10), dim = c(2, 5))
stick_breaking(Z)

}
